backend/requirements.txt
setuptools<81
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-allauth==0.57.0
dj-rest-auth==5.0.2
django-environ==0.11.2
psycopg2-binary==2.9.9
Pillow==10.1.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8